ecom-api/requirements.txt
alembic==1.8.1
astroid==2.12.12
autopep8==2.0.0
bcrypt==4.0.1
click==8.1.3
dill==0.3.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.3
psycopg2-binary==2.9.5
pycodestyle==2.9.1
PyJWT==2.6.0
pylint==2.15.5
pyparsing==3.0.9
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.43
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
